pyproject.toml
[project]
name = "codespace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "koyeb-sdk>=1.2.1",
    "python-dotenv>=1.2.1",
]
